Replace the PasswordEdit field by a changePassword QPushButton, remove foreach uses
Summary:
BUG: 328664

The current behavior to change password is quite unfriendly to say the least.
Clicking in a line field to open a dialog is not logical.
Replace this with a much simpler button leading to the dialog.

Test Plan:
Compile
kcmshell5 user_manager
Tried changing your password
Tried adding a user

Before:
{F7269649}

After:

{F7269690}

Reviewers: #plasma, ngraham, davidedmundson, sitter, #vdg, GB_2

Reviewed By: #plasma, ngraham, davidedmundson, #vdg, GB_2

Subscribers: nicolasfella, GB_2, plasma-devel

Tags: #plasma, #vdg

Differential Revision: https://phabricator.kde.org/D23337